backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.13.1
pydantic_core==2.46.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8.3
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
